Move pyramid tests to their own file
