+++ builder/test/BUILD
@@ -78,6 +78,26 @@ cc_test(
 )
 
 cc_test(
+    name = "pyramid",
+    srcs = [
+        "pyramid.cc", 
+        "bazel_util.h",
+        "funcs.h",
+        "util.h",
+    ],
+    data = ["replica_pipeline.cc"],
+    deps = [
+        "//builder",
+        "//builder:replica_pipeline",
+        "//runtime",
+        "//runtime:thread_pool",
+        "@bazel_tools//tools/cpp/runfiles",
+        "@googletest//:gtest_main",
+    ],
+    size="small",
+)
+
+cc_test(
     name = "aligned_producer",
     srcs = [
         "aligned_producer.cc",
